fix(wallets): avoid undefined fields (linkedCardId/note)
