+++ functions/tsconfig.json
@@ -1,11 +1,15 @@
 {
   "compilerOptions": {
     "target": "ES2020",
+    "lib": ["ES2020"],
     "module": "commonjs",
+    "moduleResolution": "node",
     "rootDir": "src",
     "outDir": "lib",
+    "strict": true,
     "esModuleInterop": true,
-    "strict": true
+    "skipLibCheck": true,
+    "resolveJsonModule": true
   },
-  "include": ["src"]
+  "include": ["src/**/*"]
 }
